requirements.txt
alembic==1.6.5
appdirs==1.4.4
asgiref==3.3.4
black==21.6b0
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
distlib==0.3.2
fastapi==0.65.2
FastAPI-SQLAlchemy==0.2.1
filelock==3.0.12
greenlet==2.0.2
h11==0.12.0
Mako==1.1.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
pipenv==2021.5.29
psycopg2-binary==2.9.6
pydantic==1.8.2
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
regex==2021.4.4
six==1.16.0
SQLAlchemy==1.4.19
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
email-validator==1.1.3
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
autopep8==1.5.7
bcrypt==3.2.0
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
dnspython==2.1.0
ecdsa==0.17.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
httptools==0.3.0
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
orjson==3.6.1
passlib==1.7.4
promise==2.3
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
Rx==1.6.1
ujson==4.0.2
urllib3==1.26.6
websockets==8.1
watchdog==2.1.9